weight into its own module and assert in release
